beta_mvp/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
celery>=5.3.0
redis>=5.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
google-generativeai>=0.8.0
python-telegram-bot>=20.0
discord.py>=2.3.0
faiss-cpu>=1.7.4
alembic>=1.13.0
python-dotenv>=1.0.0
